Handle file errors
This handles the cases in which a file is not given, doesn't exist, or
is empty.

The script is now callable with an argument.

Added an empty file for test purposes, but given more time would have
liked to have created this inside the test and then destroyed it after.